+++ mirrord-layer/src/error.rs
@@ -1,4 +1,4 @@
-use std::{env::VarError, net::SocketAddr, os::unix::io::RawFd, ptr, str::ParseBoolError};
+use std::{env::VarError, os::unix::io::RawFd, ptr, str::ParseBoolError};
 
 use errno::set_errno;
 use kube::config::InferConfigError;