+++ tests/src/sanity.rs
@@ -751,23 +751,6 @@ mod tests {
         process.assert_stderr();
     }
 
-    #[rstest]
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    #[timeout(Duration::from_secs(240))]
-    pub async fn test_remote_dns_lookup_pod_service(#[future] service: EchoService) {
-        let service = service.await;
-        let node_command = vec![
-            "node",
-            "node-e2e/remote_dns/test_remote_dns_lookup_pod_service.mjs",
-        ];
-        let mirrord_args = vec!["-d"];
-        let mut process = run(node_command, &service.pod_name, None, Some(mirrord_args)).await;
-
-        let res = process.child.wait().await.unwrap();
-        assert!(res.success());
-        process.assert_stderr();
-    }
-
     #[cfg(target_os = "linux")]
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]